# packages/api/swagger.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:8080","description":"Local Server"}],"info":{"version":"0.5.16","title":"akcse-mb-server","description":"","license":{"name":"UNLICENSED"}},"paths":{"/v1/account/tokens":{"post":{"summary":"create token for user","tags":["account"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoTokenDTO"}}}}}}},"/v1/configuration":{"get":{"summary":"get server configuration","tags":["configuration"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerConfig"}}}}}}},"/v1/configuration/healthcheck":{"get":{"summary":"get server status","tags":["configuration"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerMetaData"}}}}}}},"/v1/event":{"get":{"summary":"get events","tags":["event"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoGetEventsOutput"}}}}}},"post":{"summary":"create a new event","tags":["event"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCreateRequestDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject"}}}}},"security":[{"bearer":[]}]}},"/v1/event/past":{"get":{"summary":"get past events","tags":["event"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoArrayEventResponseDTO"}}}}}}},"/v1/event/{id}":{"get":{"summary":"get an event with matching ID","tags":["event"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoEventResponseDTO"}}}}}},"put":{"summary":"update an existing event's details","tags":["event"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpdateRequestDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject"}}}}},"security":[{"bearer":[]}]},"delete":{"summary":"Delete an event","tags":["event"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject"}}}}},"security":[{"bearer":[]}]}},"/v1/members":{"post":{"summary":"create a new member","tags":["members"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCreateRequestDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o1"}}}}},"security":[{"bearer":[]}]},"get":{"summary":"get all registered members","tags":["members"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoArrayMembersResponseDTO"}}}}}}},"/v1/members/leaderboard":{"get":{"summary":"get top 5 members with the highest scores","tags":["members"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoArrayTopMembersResponseDTO"}}}}}}},"/v1/members/admin":{"get":{"summary":"get all admins","tags":["members"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoArrayAdminMembersResponseDTO"}}}}}}},"/v1/members/{id}":{"get":{"summary":"get a member with the passed id","tags":["members"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoMembersResponseDTO"}}}}}},"put":{"summary":"Update an existing member's details","tags":["members"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberUpdateRequestDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o1"}}}}},"security":[{"bearer":[]}]},"delete":{"summary":"Delete a member","tags":["members"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o1"}}}}},"security":[{"bearer":[]}]}},"/v1/resources":{"post":{"summary":"create a new resource","tags":["resources"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourcesCreateDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o2"}}}}},"security":[{"bearer":[]}]},"get":{"summary":"get all posted resources","tags":["resources"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoArrayResourceResponseDTO"}}}}}}},"/v1/resources/{id}":{"get":{"summary":"get a resource with the passed id","tags":["resources"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoResourceResponseDTO"}}}}}},"put":{"summary":"Update an existing resource's details","tags":["resources"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceUpdateDTO"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o2"}}}}},"security":[{"bearer":[]}]},"delete":{"summary":"Delete a resource","tags":["resources"],"parameters":[{"name":"id","in":"path","schema":{"type":"number"},"required":true}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseResponseDtoobject.o2"}}}}},"security":[{"bearer":[]}]}}},"components":{"schemas":{"BaseResponseDtoTokenDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/TokenDTO"}},"required":["data"]},"TokenDTO":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"accessTokenExpiredAt":{"type":"string","format":"date-time"},"refreshTokenExpiredAt":{"type":"string","format":"date-time"}},"required":["accessToken","refreshToken","accessTokenExpiredAt","refreshTokenExpiredAt"]},"CreateTokenRequest":{"type":"object","properties":{"code":{"type":"string"}},"required":["code"]},"ServerConfig":{"type":"object","properties":{"env":{"oneOf":[{"const":"local"},{"const":"test"},{"const":"dev"},{"const":"prod"}]}},"required":["env"]},"ServerMetaData":{"type":"object","properties":{"serverTime":{"type":"string"},"version":{"type":"string"}},"required":["serverTime","version"]},"BaseResponseDtoGetEventsOutput":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/GetEventsOutput"}},"required":["data"]},"GetEventsOutput":{"type":"object","properties":{"upcoming":{"type":"array","items":{"$ref":"#/components/schemas/EventResponseDTO"}},"past":{"type":"array","items":{"$ref":"#/components/schemas/EventResponseDTO"}}},"required":["upcoming","past"]},"EventResponseDTO":{"type":"object","properties":{"id":{"type":"number","description":"event id\nexample: 0"},"title":{"type":"string","description":"event title\nexample: \"title\""},"description":{"type":"string","description":"event description\nexample: \"description\""},"fee":{"type":"number","description":"event fee\nexample: 0"},"startDateTime":{"type":"string","description":"event start date time\nexample: \"2024-12-27T00:00:00Z\"","format":"date-time"},"endDateTime":{"type":"string","description":"event end date time\nexample: \"2024-12-27T00:00:00Z\"","format":"date-time"},"location":{"type":"string","description":"event location\nexample: \"location\""},"signUpDeadline":{"type":"string","description":"event sign up deadline\nexample: \"2024-12-27T00:00:00Z\"","format":"date-time"},"createdAt":{"type":"string","description":"event created at\nexample: \"2024-12-27T00:00:00Z\"","format":"date-time"},"updatedAt":{"type":"string","description":"event updated at\nexample: \"2024-12-27T00:00:00Z\"","format":"date-time"},"rsvpLink":{"type":"string","description":"event rsvp link\nexample: \"rsvpLink\""},"imageUrl":{"type":"string","description":"event image url\nexample: \"imageUrl\""}},"required":["id","title","description","fee","startDateTime","endDateTime","location","signUpDeadline","createdAt","updatedAt","rsvpLink","imageUrl"]},"BaseResponseDtoArrayEventResponseDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/EventResponseDTO"}}},"required":["data"]},"BaseResponseDtoEventResponseDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/EventResponseDTO"}},"required":["data"]},"BaseResponseDtoobject":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/object"}},"required":["data"]},"object":{"type":"object","properties":{}},"EventCreateRequestDTO":{"type":"object","properties":{"startDateTime":{"type":"string"},"endDateTime":{"type":"string"},"signUpDeadline":{"type":"string"},"title":{"type":"string","description":"event title\nexample: \"title\""},"description":{"type":"string","description":"event description\nexample: \"description\""},"fee":{"type":"number","description":"event fee\nexample: 0"},"location":{"type":"string","description":"event location\nexample: \"location\""},"rsvpLink":{"type":"string","description":"event rsvp link\nexample: \"rsvpLink\""},"imageUrl":{"type":"string","description":"event image url\nexample: \"imageUrl\""}},"required":["startDateTime","endDateTime","signUpDeadline","title","description","fee","location"]},"EventUpdateRequestDTO":{"type":"object","properties":{"startDateTime":{"type":"string"},"endDateTime":{"type":"string"},"signUpDeadline":{"type":"string"},"title":{"type":"string","description":"event title\nexample: \"title\""},"description":{"type":"string","description":"event description\nexample: \"description\""},"fee":{"type":"number","description":"event fee\nexample: 0"},"location":{"type":"string","description":"event location\nexample: \"location\""},"rsvpLink":{"type":"string","description":"event rsvp link\nexample: \"rsvpLink\""},"imageUrl":{"type":"string","description":"event image url\nexample: \"imageUrl\""}}},"BaseResponseDtoobject.o1":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/object.o1"}},"required":["data"]},"object.o1":{"type":"object","properties":{}},"MemberCreateRequestDTO":{"type":"object","properties":{"score":{"type":"number"},"name":{"type":"string"},"role":{"oneOf":[{"const":"MEMBER"},{"const":"ADMIN"},{"const":"UNKNOWN"}]}},"required":["score","name","role"]},"BaseResponseDtoArrayMembersResponseDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/MembersResponseDTO"}}},"required":["data"]},"MembersResponseDTO":{"type":"object","properties":{"id":{"type":"number"},"score":{"type":"number"},"name":{"type":"string"},"role":{"oneOf":[{"const":"MEMBER"},{"const":"ADMIN"},{"const":"UNKNOWN"}]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","score","name","role","createdAt","updatedAt"]},"BaseResponseDtoArrayTopMembersResponseDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/TopMembersResponseDTO"}}},"required":["data"]},"TopMembersResponseDTO":{"type":"object","properties":{"id":{"type":"number"},"score":{"type":"number"},"name":{"type":"string"},"role":{"oneOf":[{"const":"MEMBER"},{"const":"ADMIN"},{"const":"UNKNOWN"}]}},"required":["id","score","name","role"]},"BaseResponseDtoArrayAdminMembersResponseDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AdminMembersResponseDTO"}}},"required":["data"]},"AdminMembersResponseDTO":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"role":{"oneOf":[{"const":"MEMBER"},{"const":"ADMIN"},{"const":"UNKNOWN"}]}},"required":["id","name","role"]},"BaseResponseDtoMembersResponseDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/MembersResponseDTO"}},"required":["data"]},"MemberUpdateRequestDTO":{"type":"object","properties":{"score":{"type":"number"},"name":{"type":"string"},"role":{"oneOf":[{"const":"MEMBER"},{"const":"ADMIN"},{"const":"UNKNOWN"}]}}},"BaseResponseDtoobject.o2":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/object.o2"}},"required":["data"]},"object.o2":{"type":"object","properties":{}},"ResourcesCreateDTO":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"academicCalendarUrl":{"type":"string"}},"required":["title","description","academicCalendarUrl"]},"BaseResponseDtoArrayResourceResponseDTO":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ResourceResponseDTO"}}},"required":["data"]},"ResourceResponseDTO":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"academicCalendarUrl":{"oneOf":[{"type":"null"},{"type":"string"}]}},"required":["id","title","description","academicCalendarUrl"]},"BaseResponseDtoResourceResponseDTO":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/ResourceResponseDTO"}},"required":["data"]},"ResourceUpdateDTO":{"type":"object","properties":{"description":{"type":"string"},"academicCalendarUrl":{"type":"string"}}}},"securitySchemes":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}}},"tags":[{"name":"account"},{"name":"configuration"},{"name":"event"},{"name":"members"},{"name":"resources"}],"x-samchon-emended":true}